use a hybrid rendering approach for direct3d
this change updates RenderCanvas to allow it to selectively render primitives, defaulting to all.  this is then displayed on top of the direct3d renderer set to only render points and text (lines and ellipses are handled in 3d).  this retains the high performance of the direct3d rendering while allowing the simple rendering of points and text, which are expected to be relatively uncommon primitives.